Moved input file copy from hydra_audio into the pipeline_input directory to pytest_sessionstart() to ensure it happens only once. Will hopefully fix the mysterious quick test failure on Jenkins
