+++ test/pipeline/conftest.py
@@ -3,6 +3,7 @@
 
 import pytest
 import os, re, sys
+import subprocess
 
 hydra_audio_base_dir = os.path.expanduser("~/hydra_audio/")
 pipeline_input_dir = os.path.abspath("./pipeline_input/")
@@ -31,6 +32,32 @@ quick_test_pass_thresholds = {
 "InHouse_XVF3510v080_v1.2_20190423_Loc1_Noise1_65dB_XMOS_DUT1_80dB_Take1.wav" : 24, #24 max score. AEC and IC test
 "InHouse_XVF3510v080_v1.2_20190423_Loc2_Noise1_65dB__Take1.wav" : 24, #25 max score. IC test mainly
 }
+
+def get_wav_info(input_file):
+    chans = int(subprocess.check_output(("soxi", "-c", input_file)))
+    rate = int(subprocess.check_output(("soxi", "-r", input_file)))
+    samps = int(subprocess.check_output(("soxi", "-s", input_file)))
+    bits = int(subprocess.check_output(("soxi", "-b", input_file)))
+    return chans, rate, samps, bits
+
+
+def convert_input_wav(input_file, output_file):
+    chans, rate, samps, bits = get_wav_info(input_file)
+    extra_args = "" #"trim 0 5" #to test with short wavs
+    if chans == 6:
+        #for 6 channel wav file, first 2 channels are the mic input, followed by 2 channels of far-end audio, followed by 2 channels of pipeline output
+        subprocess.run(f"sox {input_file} -r 16000 -b 32 {output_file} remix 1 2 3 4 {extra_args}".split())
+    elif chans == 8:
+        # for 8 channel wav file, first 2 channels are comms and asr outputs, followed by 4 channels of mic input
+        # and last 2 channels are far-end audio
+        subprocess.run(f"sox {input_file} -r 16000 -b 32 {output_file} remix 3 6 7 8 {extra_args}".split())
+    elif chans == 4:
+        # for 4 channel wav file, first 2 channels are mic, followed by 2 channels of reference input
+        # and last 2 channels are far-end audio
+        subprocess.run(f"sox {input_file} -r 16000 -b 32 {output_file} {extra_args}".split())
+    else:
+        assert False, f"Error: input wav format not supported - chans:{chans}"
+    return output_file
 
 # This is a list of tuples we will build consisting of test_wav and target
 all_tests_list = []
@@ -87,6 +114,13 @@ def pytest_sessionstart(session):
         for arch in architectures:
             os.makedirs(os.path.join(pipeline_output_base_dir+"_"+arch+"_"+target), exist_ok=True)
             os.makedirs(os.path.join(keyword_input_base_dir+"_"+arch+"_"+target), exist_ok=True)
+    
+    for test in all_tests_list:
+        wav_file = test[0]
+        wav_name = os.path.basename(wav_file)
+        input_file = os.path.join(pipeline_input_dir, wav_name)
+        print(f"convert_file: {wav_file}, {input_file}")
+        convert_input_wav(wav_file, input_file)
 
     #Start with empty logfile
     open(results_log_file, "w").close()
